Rename the central file for loading MiniPortile to mini_portile2.rb
This makes sure that we do not break the installation of already released
gems of other authors. And in conjunction with the clear recommendation to
place a gem version constraint into the other gem's extconf.rb file, it
makes it safe for us to change MiniPortile's behavior in the future,
without breaking their installation.

See here for the discussion:
  https://groups.google.com/d/msg/nokogiri-talk/FLm2_U0KesU/7Y8iutkCEwAJ