Fix GPG verification on Windows and add GPG2
1. Use IO.popen to avoid issues with missing escaping in the paths.
2. Remove unused method normalize_path.
3. Try both gpg2 and gpg for signature verification.
4. Use relative path for keyring and delete the key afterwards.

To point 4:

The reason for use relative paths is that there are 3 different flavors of gpg
are used on Windows, which differ in the way how absolute paths are specified:

MINGW: C:\path\to\keyring.gpg
MSYS2: /c/path/to/keyring.gpg
Cygwin: /cygdrive/c/path/to/keyring.gpg

Since there is no reliable way to find out which flavor of gpg is actually
running, it's best to use relative paths only. The downside is, that we can
not reliably say where the file is located in the end, which is why only the
imported key is deleted afterwards, not the whole keyring file.