Explicitly use GNU make for FreeBSD build
By default, "make" on FreeBSD is not compatible with the GNU make
expected by the example libraries that test:examples will attempt to
compile.

GNU make _is_ available for FreeBSD but isn't installed by default so we
explicitly download it and set the MAKE environment variable so
MiniPortile will use it.